separated typedef test into its own file, working on education testing
